requirements.txt
pykrx
pandas
gspread
oauth2client
python-dateutil
